* V5-7-patches:
  CHANGES: libnetsnmp: PATCH 3536692: from mceier: create a unit test for read_config APIs
  CHANGES: libnetsnmp: PATCH 3536692: from mceier: Fix for callback calling when unregistering config handler
  don't duplicate send_enterprise_trap_vars def
  CHANGES: freebsd: PATCH 3383970: from yaberauneya: Fix compile error on FreeBSD 9.x/POSIXify auto_nlist, etc

+++ agent/kernel.h
@@ -17,7 +17,7 @@
 
 #ifdef NETSNMP_CAN_USE_NLIST
 extern int      init_kmem(const char *);
-extern int      klookup(unsigned long, char *, int);
+extern int      klookup(unsigned long, char *, size_t);
 extern void     free_kmem(void);
 #define NETSNMP_KLOOKUP(x,y,z) klookup((unsigned long) x,y,z)
 #else
